HSEARCH-4132 Move the SerializationUtils to common
